Langeweile

+++ core-lib/src/main/java/net/verany/api/player/permission/duration/GroupTime.java
@@ -9,7 +9,7 @@ public class GroupTime extends AbstractGroupTime {
     public static final AbstractGroupTime MONTHS = new GroupTime(TimeUnit.DAYS, 30, "months");
     public static final AbstractGroupTime HOURS = new GroupTime(TimeUnit.HOURS, "hours");
     public static final AbstractGroupTime DAYS = new GroupTime(TimeUnit.DAYS, "days");
-    public static final GroupDuration LIFETIME = new GroupDuration(-1, Integer.MAX_VALUE, "lifetime");
+    public static final GroupDuration LIFETIME = new GroupDuration(-1, Long.MAX_VALUE, "lifetime");
     public static final GroupDuration SEASON = new GroupDuration(90, System.currentTimeMillis() - Season.getCurrentSeason().getStart().getTimeInMillis(), "season");
 
     public GroupTime(TimeUnit timeUnit, String key) {
